+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Next
+- Add ability to delete tracks from library
 - Add `Escape` shortcut for deselecting tracks
 - Add support selecting tracks via `Cmd+A`, `Shift+Up/Down`, `Alt+Up/Down`, `Shift+Alt+Up/Down`
 - Improve behavior of shift-selecting tracks